Move TransmuterSymbolMatchError to common library logic, check if next_terminal returns None when starting parsing
