+++ lib/Python/transmuter/front/common.py
@@ -32,3 +32,13 @@ class Position:
 class TransmuterException(Exception):
     def __init__(self, filename: str, position: Position, type_: str, description: str) -> None:
         super().__init__(f"{filename}:{position.line}:{position.column}: {type_}: {description}")
+
+
+class TransmuterInternalError(TransmuterException):
+    def __init__(self, description: str) -> None:
+        super().__init__("<internal>", Position(0, 0, 0), "Internal Error", description)
+
+
+class TransmuterSymbolMatchError(TransmuterInternalError):
+    def __init__(self) -> None:
+        super().__init__("Could not match symbol.")
